feat: duration

+++ packages/nocodb-sdk/src/lib/sqlUi/MssqlUi.ts
@@ -1147,7 +1147,7 @@ export class MssqlUi {
         colProp.dt = 'double';
         break;
       case 'Duration':
-        colProp.dt = 'int';
+        colProp.dt = 'decimal';
         break;
       case 'Rating':
         colProp.dt = 'int';
